Refactored code and FE fixes and improvements on the data browser extensions
