+++ src/app/data/components/data-browser/data-browser.component.ts
@@ -55,7 +55,8 @@ import { DownloadState } from 'src/app/import/services/s3.enums';
 import { UserManager } from 'src/app/util/user-manager';
 import { labelingLinkType } from 'src/app/labeling/components/helper/labeling-helper';
 import { CommentDataManager, CommentType } from 'src/app/base/components/comment/comment-helper';
-import { getSearchOperatorTooltip, SearchOperator } from './helper-classes/search-operators';
+import { getAttributeType, getSearchOperatorTooltip, SearchOperator } from './helper-classes/search-operators';
+import { UpdateSearchParameters } from './helper-classes/update-search-parameters';
 
 
 type DataSlice = {
@@ -192,6 +193,7 @@ export class DataBrowserComponent implements OnInit, OnDestroy {
   saveDropdonwAttribute: string = "";
   colorsAttributes: string[] = [];
   separator: string = ",";
+  updateSearchParameters: UpdateSearchParameters;
 
   getSearchFormArray(groupKey: string): FormArray {
     return this.fullSearch.get(groupKey).get('groupElements') as FormArray;
@@ -229,6 +231,7 @@ export class DataBrowserComponent implements OnInit, OnDestroy {
     this.refreshAnyRecordManuallyLabeled(this.projectId);
     this.filterParser = new DataBrowserFilterParser(this);
     this.userFilter = new UserFilter(this, this.organizationApolloService);
+    this.updateSearchParameters = new UpdateSearchParameters(this);
 
     let preparationTasks$ = [];
     preparationTasks$.push(this.userFilter.prepareUserRequest());
@@ -398,7 +401,7 @@ export class DataBrowserComponent implements OnInit, OnDestroy {
     );
     //Drill Down
     const group = this.formBuilder.group({ DRILL_DOWN: false });
-    group.valueChanges.subscribe(() => this.refreshSearchParamText());
+    group.valueChanges.subscribe(() => this.updateSearchParameters.refreshSearchParamText());
     this.fullSearch.set("DRILL_DOWN", group);
 
     //attributes
@@ -526,7 +529,7 @@ export class DataBrowserComponent implements OnInit, OnDestroy {
     this.onlyKeepUpdatedGroup(group, updatedValues);
     values = group.getRawValue();
     values.active = this.anyOrderActive(values);
-    this.refreshSearchParams(values);
+    this.updateSearchParameters.refreshSearchParams(values);
     if (!values.active) this.staticSliceOrderActive = "";
     if (this.activeSlice?.static && this.activeSlice?.sliceType != this.SliceTypes.STATIC_OUTLIER) {
       this.requestExtendedSearchByStaticSlice();
@@ -541,22 +544,7 @@ export class DataBrowserComponent implements OnInit, OnDestroy {
     }
     return false;
   }
-  private _orderByBuildSearchParamText(values): string {
-    let text = '';
-    for (const element of values.orderBy) {
-      if (element.active) {
-        if (text) text += "\n";
-        else text = "ORDER BY "
-        text += element.displayName;
-        if (element.displayName != this._getOrderByDisplayName(StaticOrderByKeys.RANDOM)) {
-          text += (element.direction == 1 ? ' ASC' : ' DESC');
-        } else {
-          text += ' (seed:' + element.seedString + ')';
-        }
-      }
-    }
-    return text;
-  }
+
 
   public onlyKeepUpdatedGroup(formItem: FormGroup | FormArray | FormControl, updatedValues: any,
     name?: string) {
@@ -700,14 +688,7 @@ export class DataBrowserComponent implements OnInit, OnDestroy {
 
   }
 
-  private _getOrderByDisplayName(orderByKey: string) {
-    switch (orderByKey) {
-      case StaticOrderByKeys.RANDOM: return "Random";
-      case StaticOrderByKeys.WEAK_SUPERVISION_CONFIDENCE: return "Weak Supervision Confidence";
-      case StaticOrderByKeys.MODEL_CALLBACK_CONFIDENCE: return "Model Callback Confidence";
-      default: return orderByKey; //attributes
-    }
-  }
+
 
   getOrderByGroup(orderByKey: string, isAttribute: boolean, direction) {
     if (!this.allOrderByGroups.has(orderByKey)) {
@@ -718,7 +699,7 @@ export class DataBrowserComponent implements OnInit, OnDestroy {
           orderByKey: orderByKey,
           active: false,
           seedString: "",
-          displayName: this._getOrderByDisplayName(orderByKey),
+          displayName: this.getOrderByDisplayName(orderByKey),
           isAttribute: isAttribute,
         });
         this.groupValueChangesSubscribtion$.push(group.valueChanges.subscribe(() => {
@@ -734,7 +715,7 @@ export class DataBrowserComponent implements OnInit, OnDestroy {
           orderByKey: orderByKey,
           active: false,
           direction: direction,
-          displayName: this._getOrderByDisplayName(orderByKey),
+          displayName: this.getOrderByDisplayName(orderByKey),
           isAttribute: isAttribute,
         });
       }
@@ -826,7 +807,7 @@ export class DataBrowserComponent implements OnInit, OnDestroy {
       group.get('active').setValue(false);
       return;
     }
-    this.refreshSearchParams(values);
+    this.updateSearchParameters.refreshSearchParams(values);
     this.checkAndDisplayDisplayValuesChangedWarning();
     if (this.activeSlice?.static) {
       this.checkFilterChangedForStaticSlice();
@@ -986,7 +967,7 @@ export class DataBrowserComponent implements OnInit, OnDestroy {
       group.get('active').setValue(true, { emitEvent: false });
       values.active = true;
     }
-    this.refreshSearchParams(values);
+    this.updateSearchParameters.refreshSearchParams(values);
     this.checkAndDisplayDisplayValuesChangedWarning();
     if (this.activeSlice?.static) {
       this.refreshHighlightModule();
@@ -1002,121 +983,122 @@ export class DataBrowserComponent implements OnInit, OnDestroy {
     return true;
   }
 
-  refreshSearchParamText() {
-    for (let p of this.activeSearchParams) {
-      this.updateSearchParamText(p);
-      this.createSplittedText(p);
-    }
-  }
-  createSplittedText(p) {
-    const groupName = this.searchGroups.get(p.values.groupKey).nameAdd + ':';
-    p.searchTextReplaced = p.searchText.replaceAll("\nAND", "\n<gn>" + groupName + "\n");
-    p.splittedText = p.searchTextReplaced.split("\n<gn>");
-  }
+  // refreshSearchParamText() {
+  //   for (let p of this.activeSearchParams) {
+  //     this.updateSearchParamText(p);
+  //     this.createSplittedText(p);
+  //   }
+  // }
+  // createSplittedText(p) {
+  //   const groupName = this.searchGroups.get(p.values.groupKey).nameAdd + ':';
+  //   p.searchTextReplaced = p.searchText.replaceAll("\nAND", "\n<gn>" + groupName + "\n");
+  //   p.splittedText = p.searchTextReplaced.split("\n<gn>");
+  // }
 
-  updateSearchParam(searchElement, newValues) {
-    searchElement.values = newValues;
-    this.updateSearchParamText(searchElement);
-    this.createSplittedText(searchElement);
-  }
+  // updateSearchParam(searchElement, newValues) {
+  //   searchElement.values = newValues;
+  //   this.updateSearchParamText(searchElement);
+  //   this.createSplittedText(searchElement);
+  // }
 
-  updateSearchParamText(searchElement) {
-    if (searchElement.values.type == SearchItemType.ATTRIBUTE) {
-      const attributeType = this.getAttributeType(searchElement.values.name);
-      if (searchElement.values.operator == SearchOperator.BETWEEN) {
-        if (attributeType == "INTEGER" || attributeType == "FLOAT") {
-          searchElement.searchText =
-            searchElement.values.name +
-            ' ' +
-            searchElement.values.operator +
-            " " +
-            searchElement.values.searchValue +
-            "" + " AND " + searchElement.values.searchValueBetween;
-        } else {
-          searchElement.searchText =
-            searchElement.values.name +
-            ' ' +
-            searchElement.values.operator +
-            " '" +
-            searchElement.values.searchValue +
-            "'" + " AND '" + searchElement.values.searchValueBetween + "'";
-        }
-      } else if (searchElement.values.operator == '') {
-        searchElement.searchText = searchElement.values.name;
-      } else if (searchElement.values.operator == SearchOperator.IN) {
-        if (attributeType == "INTEGER" || attributeType == "FLOAT") {
-          searchElement.searchText =
-            searchElement.values.name +
-            ' ' +
-            searchElement.values.operator +
-            " (" +
-            searchElement.values.searchValue + ")";
-        } else {
-          const splitTextBySeparator = searchElement.values.searchValue.split(this.separator);
-          searchElement.searchText = searchElement.values.name + ' ' + searchElement.values.operator + " (" + splitTextBySeparator.map(x => "'" + x + "'").join(", ") + ")";
-        }
-      }
-      else {
-        if (attributeType == "INTEGER" || attributeType == "FLOAT") {
-          searchElement.searchText =
-            searchElement.values.name +
-            ' ' +
-            searchElement.values.operator +
-            " " +
-            searchElement.values.searchValue;
-        }
-        else {
-          searchElement.searchText =
-            searchElement.values.name +
-            ' ' +
-            searchElement.values.operator +
-            " '" +
-            searchElement.values.searchValue +
-            "'";
-        }
-      }
-      if (searchElement.values.negate)
-        searchElement.searchText = 'NOT (' + searchElement.searchText + ')';
-      if (this.separator == "-")
-        searchElement.searchText = searchElement.searchText.replaceAll("-", ",");
-    } else if (searchElement.values.type == SearchItemType.LABELING_TASK) {
-      searchElement.searchText = this._labelingTaskBuildSearchParamText(
-        searchElement.values
-      );
-    } else if (searchElement.values.type == SearchItemType.USER) {
-      searchElement.searchText = this.userFilter.buildSearchParamText(
-        searchElement.values
-      );
-    } else if (searchElement.values.type == SearchItemType.ORDER_BY) {
-      searchElement.searchText = this._orderByBuildSearchParamText(
-        searchElement.values
-      );
+  // updateSearchParamText(searchElement) {
+  //   if (searchElement.values.type == SearchItemType.ATTRIBUTE) {
+  //     const attributeType = this.getAttributeType(searchElement.values.name);
+  //     if (searchElement.values.operator == SearchOperator.BETWEEN) {
+  //       if (attributeType == "INTEGER" || attributeType == "FLOAT") {
+  //         searchElement.searchText =
+  //           searchElement.values.name +
+  //           ' ' +
+  //           searchElement.values.operator +
+  //           " " +
+  //           searchElement.values.searchValue +
+  //           "" + " AND " + searchElement.values.searchValueBetween;
+  //       } else {
+  //         searchElement.searchText =
+  //           searchElement.values.name +
+  //           ' ' +
+  //           searchElement.values.operator +
+  //           " '" +
+  //           searchElement.values.searchValue +
+  //           "'" + " AND '" + searchElement.values.searchValueBetween + "'";
+  //       }
+  //     } else if (searchElement.values.operator == '') {
+  //       searchElement.searchText = searchElement.values.name;
+  //     } else if (searchElement.values.operator == SearchOperator.IN) {
+  //       if (attributeType == "INTEGER" || attributeType == "FLOAT") {
+  //         searchElement.searchText =
+  //           searchElement.values.name +
+  //           ' ' +
+  //           searchElement.values.operator +
+  //           " (" +
+  //           searchElement.values.searchValue + ")";
+  //       } else {
+  //         const splitTextBySeparator = searchElement.values.searchValue.split(this.separator);
+  //         searchElement.searchText = searchElement.values.name + ' ' + searchElement.values.operator + " (" + splitTextBySeparator.map(x => "'" + x + "'").join(", ") + ")";
+  //       }
+  //     }
+  //     else {
+  //       if (attributeType == "INTEGER" || attributeType == "FLOAT") {
+  //         searchElement.searchText =
+  //           searchElement.values.name +
+  //           ' ' +
+  //           searchElement.values.operator +
+  //           " " +
+  //           searchElement.values.searchValue;
+  //       }
+  //       else {
+  //         searchElement.searchText =
+  //           searchElement.values.name +
+  //           ' ' +
+  //           searchElement.values.operator +
+  //           " '" +
+  //           searchElement.values.searchValue +
+  //           "'";
+  //       }
+  //     }
+  //     if (searchElement.values.negate)
+  //       searchElement.searchText = 'NOT (' + searchElement.searchText + ')';
+  //     if (this.separator == "-")
+  //       searchElement.searchText = searchElement.searchText.replaceAll("-", ",");
+  //   } else if (searchElement.values.type == SearchItemType.LABELING_TASK) {
+  //     searchElement.searchText = this._labelingTaskBuildSearchParamText(
+  //       searchElement.values
+  //     );
+  //   } else if (searchElement.values.type == SearchItemType.USER) {
+  //     searchElement.searchText = this.userFilter.buildSearchParamText(
+  //       searchElement.values
+  //     );
+  //   } else if (searchElement.values.type == SearchItemType.ORDER_BY) {
+  //     searchElement.searchText = this._orderByBuildSearchParamText(
+  //       searchElement.values
+  //     );
 
-      this.staticSliceOrderActive = searchElement.searchText.replace("ORDER BY ", "");
-    }
-  }
-  refreshSearchParams(values) {
-    for (let p of this.activeSearchParams) {
-      if (p.id == values.id) {
-        if (values.active) {
-          p.values = values;
-          this.updateSearchParam(p, values);
-          return;
-        } else {
-          this.activeSearchParams = this.activeSearchParams.filter(
-            (e) => e.id != values.id
-          );
-          return;
-        }
-      }
-    }
-    //doesn't exist yet
-    if (values.active) {
-      let p = { id: values.id };
-      this.updateSearchParam(p, values);
-      this.activeSearchParams.push(p);
-    }
-  }
+  //     this.staticSliceOrderActive = searchElement.searchText.replace("ORDER BY ", "");
+  //   }
+  // }
+
+  // refreshSearchParams(values) {
+  //   for (let p of this.activeSearchParams) {
+  //     if (p.id == values.id) {
+  //       if (values.active) {
+  //         p.values = values;
+  //         this.updateSearchParam(p, values);
+  //         return;
+  //       } else {
+  //         this.activeSearchParams = this.activeSearchParams.filter(
+  //           (e) => e.id != values.id
+  //         );
+  //         return;
+  //       }
+  //     }
+  //   }
+  //   //doesn't exist yet
+  //   if (values.active) {
+  //     let p = { id: values.id };
+  //     this.updateSearchParam(p, values);
+  //     this.activeSearchParams.push(p);
+  //   }
+  // }
 
   setFilterInactive(activeSearchParam) {
     for (let group of this.getSearchFormArray(activeSearchParam.values.groupKey)
@@ -1373,11 +1355,10 @@ export class DataBrowserComponent implements OnInit, OnDestroy {
   getOperatorDropdownValues(i?: number, value?: any) {
     if (this.searchOperatorDropdownArray.length == 0) {
 
-      if (this.getAttributeType(this.saveDropdonwAttribute) !== 'BOOLEAN') {
+      if (getAttributeType(this.attributesSortOrder, this.saveDropdonwAttribute) !== 'BOOLEAN') {
         for (let t of Object.values(SearchOperator)) {
           this.searchOperatorDropdownArray.push({
-            value: t,
-            dataTip: getSearchOperatorTooltip(t)
+            value: t
           });
           this.tooltipsArray.push(getSearchOperatorTooltip(t));
         }
@@ -1725,7 +1706,7 @@ export class DataBrowserComponent implements OnInit, OnDestroy {
     const setSomething = this.applyValuesToFormGroup(filterValues, orderGroup);
     let values = orderGroup.getRawValue();
     values.active = this.anyOrderActive(values);
-    this.refreshSearchParams(values);
+    this.updateSearchParameters.refreshSearchParams(values);
     if (setSomething) {
       orderGroup.get("updateDummy").setValue(true);
       this.toggleGroupMenu(key, this.getSearchGroupsHTMLByName(key), true);
@@ -1739,7 +1720,7 @@ export class DataBrowserComponent implements OnInit, OnDestroy {
     let setSomething = false;
     setSomething = this.applyValuesToFormGroup(filterGroups[key]["groupElements"][0], taskGroup);
     const rawValues = taskGroup.getRawValue();
-    this.refreshSearchParams(rawValues);
+    this.updateSearchParameters.refreshSearchParams(rawValues);
     const activeElement = taskGroup.get("active");
     activeElement.setValue(true);
     if (!rawValues.active) activeElement.setValue(false);
@@ -1761,7 +1742,7 @@ export class DataBrowserComponent implements OnInit, OnDestroy {
       }
       setSomething = this.applyValuesToFormGroup(groupElements[i], currentGroupItem) || setSomething;
       const rawValues = currentGroupItem.getRawValue();
-      this.refreshSearchParams(rawValues);
+      this.updateSearchParameters.refreshSearchParams(rawValues);
 
       //for pairwise
       const activeElement = currentGroupItem.get("active");
@@ -1779,14 +1760,14 @@ export class DataBrowserComponent implements OnInit, OnDestroy {
     if (!this.fullSearch.has(key)) { this.displayOutdatedWarning = true; return; }
     const categoryGroup: FormGroup = this.fullSearch.get(key) as FormGroup;
     this.applyValuesToFormGroup(filterGroups[key], categoryGroup);
-    this.refreshSearchParams(categoryGroup.getRawValue());
+    this.updateSearchParameters.refreshSearchParams(categoryGroup.getRawValue());
   }
 
   processDrillDown(key: string, filterGroups: Object): void {
     if (!this.fullSearch.has(key)) { this.displayOutdatedWarning = true; return; }
     const drillDown: FormGroup = this.fullSearch.get(key) as FormGroup;
     this.applyValuesToFormGroup(filterGroups[key], drillDown);
-    this.refreshSearchParams(drillDown.getRawValue());
+    this.updateSearchParameters.refreshSearchParams(drillDown.getRawValue());
   }
 
   applyValuesToFormGroup(values: any, group: FormGroup): boolean {
@@ -1872,7 +1853,7 @@ export class DataBrowserComponent implements OnInit, OnDestroy {
     for (let [key, value] of this.fullSearch) {
       if (key.startsWith(SearchGroup.LABELING_TASKS)) {
         this.clearLabelingTaskFormGroupsHelper(value);
-        this.refreshSearchParams((value as FormGroup).getRawValue());
+        this.updateSearchParameters.refreshSearchParams((value as FormGroup).getRawValue());
       }
     }
   }
@@ -2091,6 +2072,15 @@ export class DataBrowserComponent implements OnInit, OnDestroy {
     })
   }
 
+  getOrderByDisplayName(orderByKey: string) {
+    switch (orderByKey) {
+      case StaticOrderByKeys.RANDOM: return "Random";
+      case StaticOrderByKeys.WEAK_SUPERVISION_CONFIDENCE: return "Weak Supervision Confidence";
+      case StaticOrderByKeys.MODEL_CALLBACK_CONFIDENCE: return "Model Callback Confidence";
+      default: return orderByKey; //attributes
+    }
+  }
+
   buildFullLink(route: string) {
     return window.location.protocol + '//' + window.location.host + "/app" + route;
   }
@@ -2160,7 +2150,13 @@ export class DataBrowserComponent implements OnInit, OnDestroy {
 
   selectValueDropdown(value: string, i: number, field: string, key: any) {
     this.getSearchFormArray(key).controls[i].get(field).setValue(value);
-    if (field == 'name') this.saveDropdonwAttribute = value;
+    if (field == 'name') {
+      this.saveDropdonwAttribute = value;
+      if (this.getSearchFormArray(key).controls[i].get("searchValue").value != "") {
+        this.getSearchFormArray(key).controls[i].get("searchValue").setValue("");
+        this.getSearchFormArray(key).controls[i].get("searchValueBetween").setValue("");
+      }
+    }
     this.searchOperatorDropdownArray = [];
     this.tooltipsArray = [];
     this.getOperatorDropdownValues(i, value);
@@ -2169,4 +2165,5 @@ export class DataBrowserComponent implements OnInit, OnDestroy {
   getAttributeType(attributeName: string) {
     return this.attributesSortOrder.find(att => att.name == attributeName)?.type;
   }
+
 }
